Remove nullable issue
Due to changed to the player tracker, the result of PlayerTracker#get() can be null. To avoid running the command logic with a non-existent player, an early return statement has been added.
If - for some reason - this case will happen, no message is sent to the player because the check for being online was passed earlier in the logic.

Should fix #11